requirements/base.txt
awesome-slugify==1.6.5
certifi==2018.1.18
chardet==3.0.4
construct==2.5.3
defusedxml==0.5.0
Django==2.0.2
django-allauth==0.35.0
django-appconf==1.0.2
django-compressor==2.2
django-crispy-forms==1.7.0
django-environ==0.4.4
django-redis==4.8.0
# django-haystack==2.7.0
elasticsearch==2.4.1
future==0.16.0
idna==2.6
oauthlib==2.0.6
pefile==2017.11.5
Pillow==5.0.0
psycopg2==2.7.3.2
python-ptrace==0.9.3
python3-openid==3.1.0
pytz==2017.3
rcssmin==1.0.6
redis==2.10.6
regex==2018.2.3
requests==2.18.4
requests-oauthlib==0.8.0
rjsmin==1.0.12
six==1.11.0
sqlparse==0.2.4
Unidecode==0.4.21
urllib3==1.22
argon2-cffi>=16.1.0
django-taggit==0.22.2
django-markdownx==2.0.22
django-reversion==2.0.13
beautifulsoup4==4.4.0
git+git://github.com/timbirk/django-taggit-bootstrap
git+git://github.com/timbirk/django-haystack
nltk==3.2.5
